requirements.txt
transformers>=4.34.0
torch>=2.0.0
fastapi>=0.78.0
uvicorn>=0.18.0
pydantic>=1.10.0
PyYAML>=6.0
accelerate>=0.23.0
bitsandbytes>=0.40.0
safetensors>=0.3.0
huggingface_hub>=0.18.0
hf-xet